Extend Logger error method Moved ConfigService.kt Added isClosed method and isClosed checks to setDefaultUser, clearDefaultUser, setOnline, setOffline methods
